Update support page quality statements around JCK testing
Since we're now performing JCK testing and have proven and marked
some builds as JCK compliant, update the language on the support
page to reflect that.

Note this change includes a probably incomplete update to include a
"sample" of the download box that has the jck tick mark, so
that will probably need to be fixed before actually merging.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>