JS dependency and loading tweaks (#375)
* Remove use of Underscore

The very minimal way it was being used can be replaced with native JS.

* Run page-specific JS after 'DOMContentLoaded'

This is to ensure all dependencies are loaded before their attempted use.  Internet Explorer often failed due to 'platform.js' not loading in time.

* Eliminate call to 'dist/json/config.json' (i.e. 'loadPlatformsThenData')

The current config is deployed alongside the website artifacts, so use it directly.

* Updating 'CONTRIBUTING.md' to reflect recent JS changes and 'config.json' format

* minor: adding a comment on 'require()' in 'entry.js'

* Remove unused 'loadJSON' function

* Relocate calls to 'persistUrlQuery' and 'buildMenuTwisties'

* Move global 'copyClipboard' to 'installation.js'

* Move globals 'selectLatestPlatform'+'unselectLatestPlatform' into 'releases.js'

* Bump external (CDN-based) JS dependencies to latest

* Fix extracting the page name

There might be a context root prefix (e.g. '/375' during staging)