Update installation page to document installers
Installation section covers only archive files. Updates made
to move content from the openjdk-installer/wix/readme.md
repo on to the website. That readme could be reduced to cover
only the instructions developers need to create the .msi file.

Includes IcedTea-Web install instructions, so this should be
merged when that work is available in an .msi.

Closes: #478

Signed-off-by: Sue Chaplain <sue_chaplain@uk.ibm.com>